backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_login
Pillow